Add files via upload
General Working Type

First of all, we used bluetooth hc-05 module to control the car. The bluetooth program is downloaded from Google Play Store.  Movement of the car as it is described; users send to signal thanks to software from their phones. After that, the car takes signal and it moves. Bluetooth modul uses rx tx pins to communication. Rotation is proportional to the current drawn by the motors in the wheels. If you turn to the right, you should command the left rear engine and right front engine. The rationale is the same logic as the tank.

Environmental

The car does not generate gasoline or other fuel.  The car is eco-friendly. The reason of it uses Li-Ion Battery.  (We used 3S, three series, battery). Batteries can be recycled that we used. 

Sustainability

If people make as a prototypes, it’s  a possible. On real life, not possible with this materials. Should be more powerful, more energy capacity and bigger body kits (big volume). It’s mean more money.

Economical Situation

Normally, it’s cheap with this materials but we can not apply on real life. We need more money and more powerful electrical circuit to pull up high current.